Add weather measurments controller and wire up models to start saving data to the DB

+++ database/migrations/2019_04_26_013129_create_pressures_table.php
@@ -15,7 +15,7 @@ class CreatePressuresTable extends Migration
     {
         Schema::create('pressures', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('millibars');
+            $table->decimal('millibars', 10, 6);
             $table->dateTime('measurement_time');
             $table->timestamps();
         });
